pyproject.toml
[project]
name = "llamaindex-quickstart"
version = "0.1.0"
description = "Ask Anything: Gemini + LlamaIndex quickstart project"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "google-genai>=1.27.0",
    "google-generativeai>=0.8.5",
    "html2text>=2024.2.26",
    "llama-index>=0.12.52",
    "llama-index-embeddings-google-genai>=0.2.1",
    "llama-index-llms-google-genai>=0.2.5",
    "llama-index-readers-web>=0.4.5",
    "llama-index-tools-google>=0.5.0",
    "llama-index-utils-workflow>=0.3.5",
    "pillow>=11.3.0",
    "pymupdf>=1.26.3",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "ruff>=0.12.5",
]
